requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
langchain>=0.3.0
langchain-google-genai>=1.0.0
langgraph>=0.2.0
qdrant-client>=1.11.0
networkx>=3.3
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
psycopg2-binary>=2.9.9
mem0ai>=0.1.0
ragas>=0.1.21
pytest>=8.2.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
